Stop tree search at 0-0-0 instead of 0-0-1
We also need to allow taking all pieces from a heap then, so we can
reach 0-0-0.  When at 0-0-0 the player on turn wins since this means
its opponent must have taken the last item in it's previous move.
This simplifies the code and avoids an assertion failure when calling
the program with input 0 0 0.