Add script to test a new idea to map heap configurations to integers
This works by building a table of offsets for each heap which need
simply to be added to get the integer.  This scheme is simple, fast,
and needs much less space compared to the binary search tree currently
in use.  It produces contiguous integers, i.e. a one-to-one mapping
between all heap configurations (with a known maximum for each heap)
and the integers 0..n-1.